requirements.txt
crewai
PyPDF2
python-docx
python-pptx
openpyxl
beautifulsoup4
pillow
pytesseract
markdown
pandas
lxml